// src/app/on-boarding/models/asset-type.enum.ts
export enum AssetTypeEnum {
    CashAccount = 1,
    SuperannuationAccount = 2,
    RetirementIncomeStream = 3,
    InvestmentFund = 4,

    Legacy = 7,
    Other =8,

    ZZDeprecatedCentrelinkBenefit = 100000000,
    Property = 100000001,
    DirectShares = 100000002,
    WholeofLife = 509000000,
    Endowment = 509000001,
    InsuranceBond = 509000002,
    FixedInterest = 509000003,
    Alternatives = 509000004,
}